refactor(mitm-addon): extract upstream admission owner
